Convert to int. Remove trash

+++ sources/infinite-color-engine/CoreInfiniteEngine.cpp
@@ -57,7 +57,7 @@ bool CoreInfiniteEngine::getAntiFlickeringFilterState()
 	return _smoothing->getAntiFlickeringFilterState();
 }
 
-unsigned CoreInfiniteEngine::addCustomSmoothingConfig(unsigned cfgID, int settlingTime_ms, double ledUpdateFrequency_hz, double ledUpdateDelay_fr, bool pause)
+unsigned CoreInfiniteEngine::addCustomSmoothingConfig(unsigned cfgID, int settlingTime_ms, double ledUpdateFrequency_hz, int ledUpdateDelay_fr, bool pause)
 {
 	return _smoothing->addCustomSmoothingConfig(cfgID, settlingTime_ms, ledUpdateFrequency_hz, ledUpdateDelay_fr, pause);
 }
